# backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_document_api_documents_upload_post": {
        "properties": {
          "company_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_upload_post",
        "type": "object"
      },
...
      "DocumentOut": {
        "description": "Mirrors TS interface Document.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "detected_statement_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Statement Types"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "num_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Pages"
          },
          "period_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "status",
          "num_pages",
          "error",
          "detected_statement_types",
          "period_id",
          "created_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IdentityResultOut": {
        "description": "One accounting-identity check result (mirrors validation.IdentityResult).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "difference": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difference"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "lhs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lhs"
          },
          "missing_inputs": {
            "items": {
              "type": "string"
            },
            "title": "Missing Inputs",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rhs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rhs"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tolerance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tolerance"
          }
        },
        "required": [
          "key",
          "name",
          "status",
          "lhs",
          "rhs",
          "difference",
          "tolerance",
          "missing_inputs",
          "detail"
        ],
        "title": "IdentityResultOut",
        "type": "object"
      },
...
      "RatioReportOut": {
        "description": "Mirrors TS interface RatioReport.",
        "properties": {
          "period_id": {
            "title": "Period Id",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RatioResultOut"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "period_id",
          "results"
        ],
        "title": "RatioReportOut",
        "type": "object"
      },
      "RatioResultOut": {
        "description": "Mirrors TS interface RatioResult.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "missing_inputs": {
            "items": {
              "type": "string"
            },
            "title": "Missing Inputs",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "key",
          "name",
          "category",
          "status",
          "value",
          "unit",
          "missing_inputs",
          "detail"
        ],
        "title": "RatioResultOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationReportOut": {
        "description": "Accounting-identity validation for a single period.",
        "properties": {
          "period_id": {
            "title": "Period Id",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/IdentityResultOut"
            },
            "title": "Results",
            "type": "array"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "period_id",
          "results",
          "summary"
        ],
        "title": "ValidationReportOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "finsight",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/documents/upload": {
      "post": {
        "description": "Accept a PDF upload and queue the extraction pipeline.\n\nStores the file on disk, creates a ``Document`` row, and schedules the\nextraction pipeline as a background task.\n\nReturns HTTP 202 (Accepted) immediately; the client polls\nGET /documents/{id} to watch ``status`` transition through the pipeline stages:\n    uploaded \u2192 parsing \u2192 extracting \u2192 needs_review | ready | failed",
        "operationId": "upload_document_api_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "upload"
        ]
      }
    },
...
    "/api/periods/{period_id}/ratios": {
      "get": {
        "description": "Compute all financial ratios for the given period.\n\nRatios are computed deterministically from the stored canonical line items \u2014\nnever by an LLM.  Each result carries a ``status``:\n- ``ok``          \u2014 value computed successfully\n- ``unavailable`` \u2014 one or more required inputs were missing\n- ``undefined``   \u2014 math is undefined (e.g. division by zero)\n\nPass ``market_price`` and/or ``shares_outstanding`` to unlock valuation\nratios (P/E, EV/EBITDA, etc.).  Omitting them leaves those ratios as\n``unavailable`` rather than erroring.",
        "operationId": "get_period_ratios_api_periods__period_id__ratios_get",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "title": "Period Id",
              "type": "integer"
            }
          },
          {
            "description": "Current share price in reporting currency. Required for valuation ratios.",
            "in": "query",
            "name": "market_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Current share price in reporting currency. Required for valuation ratios.",
              "title": "Market Price"
            }
          },
          {
            "description": "Shares outstanding (absolute count, not millions). Required for market-cap-based valuation ratios. Falls back to the value extracted from the income statement when absent.",
            "in": "query",
            "name": "shares_outstanding",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Shares outstanding (absolute count, not millions). Required for market-cap-based valuation ratios. Falls back to the value extracted from the income statement when absent.",
              "title": "Shares Outstanding"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatioReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Period Ratios",
        "tags": [
          "ratios"
        ]
      }
    },
    "/api/periods/{period_id}/validation": {
      "get": {
        "description": "Run deterministic accounting-identity checks for the given period.\n\nVerifies accounting equations (assets = liabilities + equity, subtotals foot to\ntotals, gross_profit = revenue - COGS, etc.) within a rounding tolerance. Each\nresult carries a ``status``:\n- ``ok``          \u2014 the identity holds within tolerance\n- ``mismatch``    \u2014 the figures do not reconcile (likely an extraction error)\n- ``unavailable`` \u2014 one or more required line items were not reported\n\nLike ratios, these are computed in pure Python, never by an LLM, so the UI can\nflag inconsistent extractions instead of trusting bad data silently.",
        "operationId": "get_period_validation_api_periods__period_id__validation_get",
        "parameters": [
          {
            "in": "path",
            "name": "period_id",
            "required": true,
            "schema": {
              "title": "Period Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Period Validation",
        "tags": [
          "ratios"
        ]
      }
    }
  }
}